cmd/egrunner: expand certain env vars present in commands
When running with -out std it's annoying to see things like $HOME not
expanded in the command itself. When running with -out json, it's even
more confusing when that command makes its way into a README where the
user is forever having to remember, "Oh yeh, GITHUB_ORG means X."

Fix that.